requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
funcsigs==1.0.2
inflection==0.3.1
more-itertools==4.2.0
pluggy==0.6.0
py==1.5.4
PyHamcrest==1.9.0
pytest==3.6.3
six==1.11.0
SQLAlchemy==1.2.9
SQLAlchemy-Utils==0.33.3
SQLAlchemy-Wrapper==1.9.1
